Initialize defaults before sourcing vars file
This commit is fairly trivail, but it should resolve Issue #24.  Moves setting
of default values into a separate function that is called before and after
sourcing the vars file.

It also includes fixes for a couple yoda conditions in set_default_vars and
fixes some whitespace issues (tabs and extra whitespace at the end of lines).