Change the set_pass function to use a temporary file
When called with the same filename as the -in and -out arguments,
openssl will zero the file before asking for a password. If the
user provides a password that is too short, or types mismatching
passwords, openssl will abort leaving the key file zeroed. This
change writes to a temporary file and only overwrites the original
key on a successful exit of openssl.