move unit tests to easyrsa-unit-test repo
This modifies the op_test.sh script to pull a working copy of the actual
unit tests from the easyrsa-unit-test repo.  This allows us to update
the unit test script outside the easy-rsa repo.

A current limitation of the way we've integrated TravisCI is that the
repo and tests go hand-in-hand.  This means that if a test is broken or
otherwise needs an update, it has to come with the PR itself.  There's
never a point of re-running a TravisCI test, since the test is part of
that repo.

Signed-off-by: Eric F Crist <ecrist@secure-computing.net>