Function set_pass() now returns 0 on success in batch-mode
If changing passwords in batch-mode, the commands set-rsa-pass and
set-ec-pass allways returned with an exit code != 0 on success.

The origin for this is the implementation of the function notice() which
allway returns an return-code != 0 if called in batch-mode.

This patch fixes the bug by adding a return 0 at the end of set_pass(),
because this is the way this behaviour is fixed in other functions. I
will provide an additional pull-request to improve the functions
notice() and warn().