Verify that $EASYRSA_EXT_DIR folder can be found
When using Windows, the current working directory is not reliably
evaluated.  This can be due to user error or possible script bugs.

This means that the variable EASYRSA may have an incorrect value,
which causes EASYRSA_EXT_DIR to point to an invalid location.

Currently, the value of EASYRSA_EXT_DIR is assumed to be correct,
without verification.

Always verify that x509-types folder can be found or error-exit.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>