Expand and improve auto-escaping; Include ($) and (")
Due to using 'sed' to create a safe SSL config, easyrsa must work-around
the way sed treats '&' ampersand.  The only way to disable sed's handling
of ampersand is to escape the ampersand (Because everybody needs sed to
behave the same demented way all the time..).

EasyRSA now allows ampersand to be used as a standard character by doing
the escaping automatically.  EasyRSA also now inserts double-quote around
the variable data which is written to the safe SSL config file.

Auto-escaping also allows ($) dollar sign and even (") double-quote to
be used. However, these last two require further steps for use.

To use a ($) dollar-sign the character MUST also be escaped in 'vars'.
This is because otherwise set_var() mangles the input data.

To use a (") double-quote the character MUST also be escaped in 'vars'.
AND it MUST be enclosed by (') single-quote for the entire field.
AND it MUST be assigned externally. ie. Do *not* use set_var().
eg. In 'vars': export EASYRSA_REQ_OU='My \"red\" bikeshed'
results in EASYRSA_REQ_OU has the value 'My "red" bikeshed'

NOTE: This does not work in Windows for (") single-quote.

Tested-on: Linux, BSD, Mac and Windows.

Closes: #587

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>