Introduce 'serialNumber' field for DN (OID 2.5.4.5)
Add a *final* layer of granularity to X509 Distinguished Name.
Only used if --req-serial="<PRINTABLE>user data" is specified.

To minimize the noise to the user by this new field, change the way
that OpenSSL is called to build a CA. Always use '-batch' mode to
build the CA certificate.

User visible change when building a CA:
* Instead of being prompted for each individual DN field, now the
  user is presented with a read-out of how the fields are currently
  set. There is now only a single confirmation that all fields are
  correct.
* If '--req-serial' is not set then it is not displayed.

PRINTABLE: a-z,A-Z,0-9, -+/=.,?:()

Closes: #462 - The original proposal and prototype code.
Supersedes: #598 - Introduce 1.organizationalUnitName

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>