build_full: Remove sign_req() subshell and do full cleanup
build-x-full uses a subshell to detect errors and cleanup the files
if an error occurs.  This does not work if Ctrl-C is pressed during
the confirmation yes/no dialog.

Replace the subshell with an indicator, $on_error_build_full_cleanup,
to force cleanup() to remove the CSR, key and certificate files when
an error occurs or user presses Ctrl-C.

Closes: #680

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>