Allow --fix-offset to create post-dated certificates
Currently, --fix-offset does not allow creating post-dated cerificates.
Instead, it rolls the start date back by one year.

The reason for this behavior is that --fix-offest was only meant to fix
a day of the year but not which year.

This patch simply removes the check for a post-dated certificate, which
allows `not-before-date` to be set upto one year in the future.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>