Prefer parameter expansion '${x:-y}' to test for "null or unset"
EasyRSA allows variables to be set to "" [null], therefore, as a
standard, the script now always tests for null or unset.

IE. Include the colon ':'.

The obvious exception to this is set_var(), which remains as test
for unset ONLY.

set_var() ONLY assigns a value if the variable is currently unset,
otherwise, it would over-write a deliberately empty value.

Also, where it helps readability, separate the expansion operator
from the expandable value, with a space.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>