Fix build-ca with --passout option
When a password to be set for the generated CA is
passed in via command line options, this can lead
to problems because the private key and
certificate generation involves two invocations
of OpenSSL/LibreSSL.

The processing of $EASYRSA_PASSOUT is here
reimplemented in shell and the read password is
stored in a variable and then passed to OpenSSL
on FD 3 when needed.

This also gets rid of the two now unnecessary
temporary file copies of the password input by the
user interactively.

Also, the -keyout option is removed from the
openssl-req invocation, because it effectively
unsets the cipher set by openssl-genpkey.
(cf. openssl asn1parse -in pki/private/ca.key)