build-ca: Revert manual CA password method to temp-files
Change the integration of the following:
* build-ca: Replace password temp-file method with file-descriptors
* commit 27870d695a324e278854146afdac5d6bdade9bba

Instead of 'replacing' the standard temp-file method, the new 'stdin'
method is offered as an alternative by using option '--ca-via-stdin'

Discussion:

Using EasyRSA temp-files means that these files can be written to
specific places. In the case of the CA password, it is recommended
to use a RAM-Disk to write CA password temp-files to.

Using heredocs to pass data via file-descriptors, in all probability,
will write the heredoc data to a temp-file managed by the shell in use.

Thus, there are temp-files written in both methods.

The difference being that, the EasyRSA managed temp-files can be written to
specific places, while the shell managed temp-files are out of EasyRSA scope.

In conclusion, both methods are offered, however, the default remains
as EasyRSA temp-files for the CA password method.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>